+++ app/(home)/(board)/community/board/page.tsx
@@ -1,4 +1,4 @@
-import CommWrapper from '@/components/Posts/CommWrapper';
+import CommWrapper from '@/components/Community/CommWrapper';
 
 const BoardPage = () => {
   return <CommWrapper />;
