+++ app/(home)/(board)/community/qna/page.tsx
@@ -1,4 +1,4 @@
-import CommWrapper from '@/components/Posts/CommWrapper';
+import CommWrapper from '@/components/Community/CommWrapper';
 
 const QnaPage = () => {
   return <CommWrapper />;
