+++ app/components/Community/CommFormat.tsx
