+++ app/components/Community/CommWrapper.tsx
@@ -6,7 +6,7 @@ import { usePathname } from 'next/navigation';
 import { useRecoilValue } from 'recoil';
 import { DocumentData } from 'firebase/firestore';
 import getAdmin from '@/utils/getAdmin';
-import CommFormat from '@/components/Posts/CommFormat';
+import CommFormat from '@/components/Community/CommFormat';
 
 const CommWrapper = () => {
   const user = useRecoilValue(authState);
